Update main.py
* Update main.py

2021-02-13
Updated the queries to add data to Days and Activities table to reflect the new database structure.
Added input validation to fileTime, where the user enters a Date
Fixed main menu not correctly doing what option the user entered
Removed misc. temporary lines in the fileTime function
Changed how the fileTime function gets the activityId. NOTE: This is still broken, it gets the activityId from whichever input the user gives based on the printed numbered list of activities, which has no relation to the activityId of an activity.

* Update main.py

2021-02-14
Fixed error where the user's activity choice when filing time had no relation to choosing which activity to file (due to activityId's not necessarily being in sequential order in the Activities table)