Update main.py
2021-02-15
Fixed bug where, when attempting to file an activity on a date where the user already filed that activity, it would create a new row. Now, it will inform the user that they've already filed time for that activity on that date, and that they will override it by continuing.
NOTE: This was required to make graphing easier, as we must select the time spent on each activity and add up the times for individual activities into one value for each activities. Now that each day can only have an activity once, we don't have to worry about a user accidentally filing an activity twice and messing with the data.